Move Metrics effect to shared project
